Move old test to legacy subdir Check in official ZUGFeRD 2.2 xml examples
