+++ tests/zugferd2/Tests/Legacy/ValidatorTest.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Easybill\ZUGFeRD2\Tests;
+namespace Easybill\ZUGFeRD2\Tests\Legacy;
 
 use Easybill\ZUGFeRD2\Validator;
 use PHPUnit\Framework\TestCase;
