+++ tests/zugferd2/Tests/Legacy/references/broken_example.xml
